+++ example/babel.config.js
@@ -12,8 +12,7 @@ module.exports = function (api) {
                 {
                     extensions: ['.tsx', '.ts', '.js', '.json'],
                     alias: {
-                        // For development, we want to alias the library to the source
-                        [pak.name]: path.join(__dirname, '..', pak.source),
+                        [pak.name]: path.join(__dirname, '..', pak.main),
                     },
                 },
             ],